requirements.txt
requests
python-twitter
pytest

